+++ src/shared/components/Modal/ConfirmModal.tsx
@@ -6,7 +6,7 @@ import { useConfirmModalStore } from '@shared/context/Modal/ConfirmModalStore'
 import { useKeyClose } from '@shared/hooks/useKeyClose/useKeyClose'
 import { useOutsideClick } from '@shared/hooks/useOutsideClick/useOutsideClick'
 import { useScrollLock } from '@shared/hooks/useScrollLock/useScrollLock'
-import { theme } from '@shared/styles/theme'
+import { theme } from '@shared/styles/emotion/theme/theme'
 
 const ConfirmModal = () => {
   const { isOpen, title, description, onConfirm, onCancel, closeConfirmModal } = useConfirmModalStore()
