+++ src/shared/styles/emotion/emotion.d.ts
@@ -1,0 +1,6 @@
+import '@emotion/react'
+import { ThemeType } from './theme/type'
+
+declare module '@emotion/react' {
+  export interface Theme extends ThemeType {}
+}
