+++ src/shared/styles/emotion/utils.ts
@@ -1,0 +1,9 @@
+'use client'
+import styled from '@emotion/styled'
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const StyledGeneric = <TProps extends {}>(Component: React.FC<TProps>, ...interpolations: any[]) => styled(
+  Component,
+)<TProps>`
+  ${interpolations}
+`
