+++ src/shared/styles/scrollbar/GlobalScrollbarStyle.tsx
@@ -1,8 +1,8 @@
 'use client'
 
-import { css,Global } from '@emotion/react'
+import { css, Global } from '@emotion/react'
 
-import { theme } from '@shared/styles/theme'
+import { theme } from '@shared/styles/emotion/theme/theme'
 
 const GlobalScrollbarStyle = () => (
   <Global
